% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bind_as_xts.R
\name{bind_as_xts}
\alias{bind_as_xts}
\title{Create an xts time series object}
\usage{
bind_as_xts(ts, forecast)
}
\arguments{
\item{ts}{A time series with class "ts"}

\item{forecast}{An object with class "forecast" created by \link[forecast]{\code{forecast}}}
}
\value{
A time series with class "xts". The series will contain four
columns, x (the original time series), fitted (the future predictions),
lower, and upper (the lower and upper bounds of an 80% prediction
interval for the predictions)
}
\description{
\code{bind_as_xts} combines a time series with the output of \link[forecast]{\code{forecast}} of the forecast package to create an \link[xts]{\code{xts}} object. This is a helper function that makes it easier to model data in the \link{\code{ts}} format and then plot the results in the \link[xts]{\code{xts}} format.
}

