% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bnames}
\alias{bnames}
\title{Baby names.}
\format{A data frame with four variables: \code{year}, \code{sex},
  \code{name}, and \code{prop}.}
\usage{
bnames
}
\description{
Full baby name data provided by the SSA. This includes all names with at
least 5 uses. This data set is a reformatted version of the \code{babynames}
data set in the \code{babynames} package.
}
\keyword{datasets}

