% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bnames2}
\alias{bnames2}
\title{bnames2}
\format{\preformatted{'data.frame':	1 obs. of  135 variables:
 $ name: chr "John"
 $ 1880: num 9655
 $ 1881: num 8769
 $ 1882: num 9557
 $ 1883: num 8894
 $ 1884: num 9388
 $ 1885: num 8756
 $ 1886: num 9026
 $ 1887: num 8110
 $ 1888: num 9248
 $ 1889: num 8548
 $ 1890: num 8502
 $ 1891: num 7681
 $ 1892: num 9039
 $ 1893: num 8049
 $ 1894: num 8238
 $ 1895: num 8321
 $ 1896: num 8139
 $ 1897: num 7550
 $ 1898: num 8159
 $ 1899: num 6990
 $ 1900: num 9830
 $ 1901: num 6899
 $ 1902: num 7907
 $ 1903: num 7609
 $ 1904: num 8108
 $ 1905: num 8060
 $ 1906: num 8264
 $ 1907: num 8983
 $ 1908: num 9342
 $ 1909: num 9591
 $ 1910: num 11450
 $ 1911: num 13446
 $ 1912: num 24587
 $ 1913: num 29326
 $ 1914: num 37946
 $ 1915: num 47576
 $ 1916: num 50047
 $ 1917: num 51851
 $ 1918: num 56558
 $ 1919: num 53534
 $ 1920: num 56910
 $ 1921: num 58231
 $ 1922: num 57275
 $ 1923: num 57467
 $ 1924: num 59062
 $ 1925: num 57188
 $ 1926: num 56112
 $ 1927: num 55945
 $ 1928: num 53744
 $ 1929: num 52099
 $ 1930: num 52425
 $ 1931: num 48397
 $ 1932: num 48597
 $ 1933: num 46217
 $ 1934: num 46734
 $ 1935: num 47490
 $ 1936: num 47655
 $ 1937: num 48737
 $ 1938: num 50504
 $ 1939: num 51021
 $ 1940: num 54760
 $ 1941: num 57509
 $ 1942: num 66026
 $ 1943: num 68913
 $ 1944: num 66603
 $ 1945: num 66116
 $ 1946: num 79245
 $ 1947: num 88318
 $ 1948: num 82827
 $ 1949: num 81094
 $ 1950: num 79413
 $ 1951: num 81500
 $ 1952: num 83071
 $ 1953: num 80064
 $ 1954: num 81124
 $ 1955: num 80018
 $ 1956: num 80758
 $ 1957: num 78983
 $ 1958: num 75973
 $ 1959: num 76427
 $ 1960: num 76123
 $ 1961: num 79910
 $ 1962: num 78447
 $ 1963: num 78644
 $ 1964: num 82540
 $ 1965: num 71568
 $ 1966: num 65054
 $ 1967: num 61637
 $ 1968: num 60921
 $ 1969: num 58792
 $ 1970: num 58541
 $ 1971: num 51520
 $ 1972: num 43179
 $ 1973: num 39174
 $ 1974: num 37654
 $ 1975: num 35102
 $ 1976: num 33987
 $ 1977: num 34232
  [list output truncated]
}}
\usage{
bnames2
}
\description{
A subset of the \code{bnames} data set.
}
\keyword{datasets}

