% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exercise.R
\name{exercise}
\alias{exercise}
\title{Open an exercise}
\usage{
exercise(number = 1)
}
\arguments{
\item{number}{The number of exercise to open.}
}
\description{
\code{exercise} opens an exercise file in your RStudio scripts pane and saves a copy of the file in your working directory.
}

