% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_babyname.R
\name{get_babyname}
\alias{get_babyname}
\title{Retrieve a name from the babynames data set}
\usage{
get_babyname(name, sex)
}
\arguments{
\item{name}{A character string. The string should begin with a capital letter.}

\item{sex}{"M" for male, "F" for female}
}
\value{
A data frame with five variables: \code{year}, \code{sex},
\code{name}, \code{n}, and \code{prop} (\code{n} divided by the total
number of applicants that year)
}
\description{
The \code{\link[babynames]{babynames}} data set provides information
on names collected by the United States Social Security
Administration. The data describes how many children were given each
name for each year since 1880, as well as the proportion of children
given the name in a given year. The data is organized by year, name
and sex and includes an entry for each name used five or more times
in a single year.
}
\details{
\code{get_babyname} returns all of the entries associated with a single name and sex.
}

