% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim_whitespace.R
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Remove trailing white spaces from a string}
\usage{
trim_whitespace(string)
}
\arguments{
\item{string}{A character string or vector of character strings}
}
\value{
A character string or vector of character strings.
}
\description{
\code{trim_whitespace} removes trailing white spaces from a character
string (or each string in a vector of character strings).
Non-trailing white spaces will be preserved.
}

